example/classifier/requirements.txt
openai>=1.0.0
python-dotenv
pydantic>=2.0.0

